TTT-10: Tests separated from production code.
