+++ src/main/ttt_main.py
@@ -1,5 +1,4 @@
-
-import ttt_game
+from src.main import ttt_game
 
 # Instantiate the game instance, which gets the initial board setup from the user
 game = ttt_game.Game()
